move gql environment to common
