+++ src/renderer/QRCode.tsx
@@ -1,29 +1,9 @@
 import React, { useEffect, useRef } from "react";
 import * as libqrcode from "qrcode";
-
-import { Environment, Network, RecordSource, Store } from "relay-runtime";
 import { graphql, QueryRenderer } from "react-relay";
+
+import environment from "../common/graphqlEnvironment";
 import { QRCodeQuery } from "./__generated__/QRCodeQuery.graphql";
-
-function fetchQuery(operation: any, variables: any) {
-  return fetch("http://localhost:8080/graphql", {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify({
-      query: operation.text,
-      variables,
-    }),
-  }).then((response) => {
-    return response.json();
-  });
-}
-
-const environment = new Environment({
-  network: Network.create(fetchQuery),
-  store: new Store(new RecordSource()),
-});
 
 function QRCode() {
   return (
